strutils: fix buffer overflow in formatBiggestFloat
Previously, strutils.formatBiggestFloat was implemented using sprintf
and as such come with a risk of buffer overflow. This was observed via
tests within disruptek/insideout.

Switch the implementation to snprintf and remove the temporary buffer
used. This solves the overflow and shave ~2 KiB from the stack size
requirement for this procedure.

Additionally, extra error checking code is added for potential failures
of snprintf.