sempass2: lift unprincipled capture restriction
Given the nested procedures `a: b: c:`, where `c` closes over a local
of `a`, `c` effectively captures from `b` and `b` captures from `a`. A
procedure usable in both a compile- and run-time context may capture
from a compile-time-only procedure, so the aforementioned should work
when `a` is a compile-time-only procedure and the others are not, but
it was forbidden in practice.

The restriction is now lifted, and the code implementing the test
restructured to be more intelligible.