macros: add `stamp` for unquoting a template body (#1426)
## Summary

Add  `stamp`  to  `std/macros` , which takes a template body and applies
it immediately. This is the same behaviour as the previous  `quote` 
that was changed after https://github.com/nim-works/nimskull/pull/1393.


## Details

After https://github.com/nim-works/nimskull/pull/1393 we have a hole in
the stdlib for  `quote` -like interpolation but with automatic binding
of symbols to reduce boilerplate. Various libraries within the ecosystem
such as  `criterion`  and  `npeg`  have found the feature useful and
extensively employ it.

Reintroduces old  `quote`  behavior as  `stamp` , without support for
custom operators and explicitly document the template-like nature of 
`stamp` 's body as well as its pitfalls.

This macro was originally written and shared to Matrix by @zerbina.

---------

Co-authored-by: zerbina <100542850+zerbina@users.noreply.github.com>
Co-authored-by: Saem Ghani <saemghani+github@gmail.com>