openssl, net: remove openssl pre-1.1.1 and libressl support
OpenSSL older than 3.0 has reached EoL and should no longer be used in
production. However, since version 1.1.1 is still widely used in LTS
distributions such as RHEL 8/Ubuntu 20.04, support for 1.1.1 is still
retained.

LibreSSL support is also removed from the wrapper, as it has not been
tested and likely no longer work with the latest versions of LibreSSL.

This commit also removes various helper procedures that bridges 1.1 APIs
and 1.0/0.9 APIs as well as "hacks" within the wrapper.

Coming with this, support for `protTLSv1` in `net` has been removed, as
this TLS version is very old and is no longer safe. Direct usage of it
has already been deprecated within OpenSSL.