mirgen: lower `finally` in `mirgen`
Instead of leaving the implementation of `finally` and the exception
stack to the code generators (or a theoretical MIR pass), `mirgen` now
does the lowering itself.

This means turning `finally` into what is effectively a
block + try/except + dispatcher, which is the same thing that `ccgflow`
already did (just with access to the original structure, which makes
the translation much simpler).

Emitting the calls to the exception runtime is also moved to `mirgen`.
In order to be able to continue raising an in-flight instruction (a
re-raise statement can only raise a *caught* exception), the
`mResumeRaising` magic is introduced. `mnkRaise` should eventually
become low-level enough for the magic is no longer necessary.