jssys: implement the exception runtime
While the `.up` field of the exception could be used - like how the C
implementation does it -, this would be unnecessarily convoluted.
Instead, the implementation simply uses a `seq`-based stack, which is
currently possible due to `seq` mapping to JavaScript arrays.

`lastJSError` is still needed for `jsgen` raise implementation, so it's
kept.