compiler: add the `--timetrace` switch
It enables or disables outputting a .json dump of the traces (they're
currently recorded regardless of whether the dump is enabled).

Due to a csources compiler bug (which is already fixed in the most
recent compiler), `CmdSwitchTextKind` cannot have more than 128
elements, and therefore the unused `seqsv2` option enum values are
removed, to make space for the new switch.