sem: instantiate `Continuation` type for `.musttail` proc
Sem is the only layer where generic types can be properly instantiate, and
therefore the return `Continuation` types are instantiated right when a
`.musttail` procedure type is constructed.

The type must not be visible to macros and the rest of sem, so it's hidden
away in a new effect-list slot of the `tyProc` type. `sempass2` is changed
such that it doesn't clobber the slot. Ultimately, this is a massive hack.