mirgen, cgen: handle `Continuation` return types
The real return types and hidden environment parameter are not visible
on the `tyProc` types, requiring special handling by both `mirgen` and
`cgen`.

`mirgen` registers uses the real return type for the result variable,
and `cgen` looks at the MIR body instead of the `PType` for the return
type. The hidden env parameter is also special case by `cgen`.