treat no return type as meaning `void` in type inference (#1497)
## Summary

During type inference part of overload resolution, a procedure type
with no return type (e.g., `proc()`) is now treated as having return
type `void`.

## Details

### Previous Behaviour

A parameter with procedure type could only match an argument with
procedure if both either have an absent return type (including an
explicit `void`) prior to inference, or not.

A parameter type such as `proc(x: int)` would never match
`proc(x: auto): auto` (where the return type is later inferred as
`void`).

### New Behavior

An absent return type is treated as meaning `void` (for both the
parameter and argument types), meaning that in the aforementioned case,
a match is possible.

### Implementation

The "are both types nil or not nil" guard in `sigmatch.procTypeRel` is
removed. Prior to relating the types, no return types are turned into
`void`, with a cached instance of the void type retrieved from the
module context.

In the `options` module, the `map` procedure accepting a non-`void`
procedure argument now needs to constrain the return type as having to
be non-`void`, otherwise the `map` overload is ambiguous with the one
having a `proc(x: T)` parameter.