mirgen: make tail-call elimination togglable
Add an option for toggling tail-call elimination in `mirgen` on
and off.

It's enabled for both the C and JS backend, but not for the VM backend,
as the latter supports native tail calls.

When the option is not enabled, no tail-call elimination is performed
in `mirgen`; eligible tail calls are only translated to `mnkTailCall`
operations.