sem: wrap expression routine bodies in `return` (#1503)
## Summary

Wrap expression routine bodies in `return` statements instead of
`result` assignments. This preserves the information that the
expression is returned directly.

## Details

The information that the expression is returned directly gets lost when
lowering it to a `result = body` statement, as those assignments cannot
be distinguished from the ones written directly by the programmer.

In addition, compared to the previous transform, the `return` is
applied to the last expression in the list (if the body is an
expression list), not the whole list.

Both `docgen` and `suggest` previously had to look for doc comments and
examples within trailing top-level assignment due to the whole body
expression being used as the assignment RHS -- these workarounds are no
longer necessary and thus removed.