lang: add the `.tailcall` calling convention (#1498)
## Summary

Add the `.tailcall` calling convention to the language. It designates
routines as being guaranteed to take over their caller's stack frame,
but only when the caller also uses the `.tailcall` calling convention.

## Details

### Design

Guaranteed tail-call elimination is an important requirement for
programs written in functional style, continuation-passing styles, or
any other style that requires deep, potentially unbounded recursion.

Implementing it efficiently via a macro is hardly possible, if at all,
making it a good candidate for being a built-in language feature.

Only enforcing tail calls for `.tailcall` calls within `.tailcall`
routines, removes the need for adapter routines at the non-`.tailcall`
to `.tailcall` edge, but it also makes the static semantics of
`.tailcall` calls callsite-dependent.

### Implementation

A description of how the portable tail-call elimination works is
provided in the `tailcallelim.rst`.

The implementation is scattered throughout multiple stages:
* `sem` and `sempass2` take care of the analysis part and synthesizing/
  instantiating the needed types and routines
* `transf` and `mirgen` perform the portable tail call elimination
* `injecthooks` handles the "trailing destructor call" detection and
  related reporting
* the code generators implement storing the parameter tuple into the
  type-erased storage (the new `mStoreParams` magic)

Whether `mirgen` performs the portable tail call elimination is
configured through the `goTailCallElim` option. When not present,
must-become-sibling-call calls are translated to the new `mnkTailCall`
MIR operation.

The VM backend uses true sibling calls, as the `mStoreParams` magic
cannot be implemented by the VM due to its strict typing requirements.

---------

Co-authored-by: Saem Ghani <saemghani+github@gmail.com>