ci: pin CMake to 3.31.6 (#1507)
## Summary

Pin CMake version to 3.31.6 for the Windows CI runner, as it's the last
version with compatibility for the CMake version PCRE needs to be
built.

## Details

GitHub updated the Windows 2022 runner image to include CMake 4.0.0,
which breaks backwards compatibility and thus makes the Windows CI jobs
depending on PCRE fail.

The problematic release:
https://github.com/actions/runner-images/releases/tag/win22%2F20250330.1