ci: fail on cancellation and update CI to ubuntu-22.04 (#1509)
## Summary
Makes sure that CI will fail if any jobs are cancelled, and update CI to
Ubuntu 22.04 to avoid brownouts.

## Details
Cancellation would cause our required check "All check passed" to be
skipped, causing GitHub to consider the check cleared and allowing
merges (https://github.com/actions/runner/issues/2566). This is
troubling as events like runner brownouts would allow PRs to skip CI and
be merged.

This PR updated the conditions for "All check passed" to ensure that a
failure is raised if any individual jobs were to be cancelled. This
change has been tested in my fork and confirmed to be working:
https://github.com/alaviss/nimskull/actions/runs/14339979291

Furthermore, while the main test runners were updated in
https://github.com/nim-works/nimskull/pull/1489 to Ubuntu 22.04, we
overlooked supporting jobs which now hits us in the form of brownouts as
GitHub phases out 20.04 runners. To avoid this in the future, this PR
makes sure that any job requiring versioned Ubuntu would follow the test
matrix, and jobs that don't will use the latest image.