config: update bootstrap compiler to 0.1.0-dev.21456 (#1513)
## Summary
Current csources could not compile  `koch`  when clang is used. Update
csources to 0.1.0-dev.21456 to solve this issue.

## Details
When csources is compiled with clang on Linux (via  `CC=clang` 
environment variable), the resulting compiler crashes with  `SIGSEGV` 
when compiling  `koch` . It was found that simply updating csources
would solve the issue, so take this chance to do so as this would also
allow future compiler code to use  `tailcall` .

Fixes https://github.com/nim-works/nimskull/issues/1512