liftdestructors: fix ill-formed code (#1524)
## Summary

Fix a latent issue with the AST `liftdestructors` produces for `ref`
assignment hooks.

## Details

AST was created that uses a `ref` value directly as the condition of an
`if`. The hook is only generated when using the C backend, and since
the code compiled down to CGIR AST without error (even though it's
wrong), the code worked correctly, as C allows pointer values being
used as conditions.

Beyond being plain wrong, this will also cease to work as soon as:
* some MIR pass relies on all `mnkIf` conditions being a boolean value
* the hook is used with the other backends