system: add the `unreachable` routine (#1529)
## Summary

Add the `unreachable` routine, plus some convenience overloads, to the
`assertions`  module.  `unreachable`  behaves the same as 
`doAssert false` ,
but it also marks the control-flow path as `.noreturn`.

## Details

Most if not all projects using `case` expressions a lot tend to require
some sort of  `unreachable` , and giving its foundational nature, it
makes
sense to have `unreachable` be part of `system`.

The implementation from `compiler/utils/idioms` is moved into the
`system/assertions` module, with some improvements made to the doc
comments. All now-unused imports of the `idioms` module in the
compiler's codebase are removed.