sem: make taking the address of `sink` parameter work (#1533)
## Summary

Fix taking the address of a `sink` parameter erroneously resulting in a
static error.

## Details

Analysis of `system.addr` and `nkAddr` used the operand's type
verbatim, which led to the creation of a `ptr(sink(...))` type for
`sink` parameter operands. Depending on usage of this type, this either
caused `typeAllowed` to report an error, or it caused some other
internal failure downstream.

The analysis of `system.addr` and `nkAddr` now skips `tySink` prior to
creating the pointer type, fixing the issue.

Fixes https://github.com/nim-works/nimskull/issues/1532 .