cli_reporter: print friendlier hook mismatch error message (#1538)
## Summary
The correct hook signature is now reported, and is explained in a clear
and concise way.

## Details
Originally an unhelpful error message, 
`signature for '=hook' must be proc[T: object](x: var T)` , was
displayed to the user if they attempted to override the default
implementation for a type.

The error message for incorrect hook signatures have changed to:
* Omit the generic parameters
* Render the  `noSideEffect`  pragma in the type signature, as hooks do
not currently taint the calling code, which could lead to the
expectations of what would have been side-effect free code, to contain
side-effects.
* State the intended meaning of `T` in plain English.
* Use the correct signature for each hook.

`semOverride`  now passes the hook's magic signature rather than the
definition, and we now render the magic's signature instead of
hardcoding the hook values.

---------

Co-authored-by: zerbina <100542850+zerbina@users.noreply.github.com>