vm: add the `opcSlice` opcode
It requires two instruction words, as it has one destination and
three operands. Creating a slice with out-of-bounds indices leads to
the creation of a slice of which access causes an error, as bound checks
and subsequent error reporting is expected to be handled by the code
generator and its runtime.