vm: implement partial support for `toOpenArray` (#1542)
## Summary

Add support for all `toOpenArray` overloads except the
`ptr UncheckedArray` one for code running in the VM (both compile-time
and standalone).

## Details

To support slices, the new built-in VM type `akOpenArray` is added,
which is implemented as a pointer + length pair and represents a
*non-owning* slice.

Adding the type means:
* handling it in type translation (`vmtypegen`)
* handling it in constant data initialization (`vmserialize` for
  compile-time code; `vmrunner` for standalone code)
* handling it in to-AST deserialization (`vmcompilerserdes`)

Compared to the other VM built-in container types (string, array, seq,
etc.), `akOpenArray` is non-owning, which means that the location
storing the openArray value being accessible doesn't imply the data
pointer being safely accessible. Read and write access has to account
for this by validating the `LocHandle` returned for a slice element.

Since the extra validation has overhead compared to seq/string/array
access, seq/string/array locations are still passed as-is to
`openArray` parameters and locals -- an `akOpenArray` value is only
constructed when storing into a field/element of `openArray` type, or
when using `toOpenArray`.

For implementing `toOpenArray`, the `opcSlice` opcode is added. The
`mChckBounds`  magic is not yet implemented, meaning that no bound
checks
are performed for `toOpenArray` for code running in the VM. Since
`opcSlice` creates an empty slice for out-of-bounds operands, the
sandbox is not compromised by the missing bound checks. 

Finally, the `knownIssue` tag is removed from various tests. The
`tests/magic/toopenarray.nim` tests is split into multiple tests in
order to be able to mark the not-yet-working parts of `toOpenArray` as
known issues, without reducing test coverage for the working parts.

Fixes https://github.com/nim-works/nimskull/issues/1541