lexopt: remove bundle special case in `remaining()`
The idea I had when this "feature" was implemented is to avoid value
synthesis in `remaining()`, making sure that only the original input is
yielded. That way, you can distinguish between `-pstuff` from
`-p -stuff` when `remaining()` is used.

However, `lexopt` was designed with the premise that an user would never
have to distinguish between different kinds of inline value
representation, since "delimited" is the only supported type of inline
parameter. As such this feature is just an error that you cannot avoid.

This commit removes the special case, and synthesize a short option
bundle from the unprocessed portions of the active bundle instead.