parseutils: make `parseBiggestFloat` non-magic (#1546)
## Summary

Make  `parseutils.parseBiggestFloat`  not rely on compiler magic,
enabling changing  `std/parseutils`  to use  `openArray[char]`  in the
future. Behaviour and interface stay the same.

## Details

There's no need for  `parseBiggestFloat`  to use compiler magic, so it's
changed to be a normal procedure, making it possible to change its
implementation and interface without touching the compiler.

The implementation is part of  `parseutils`  and dependent on the
backend:
* for the C backend, the implementation is copied verbatim from 
`strmantle.nimParseBiggestFloat` 
* for the JS backend, the implementation is copied verbatim from 
`jssys.nimParseBiggestFloat` 
* the VM implements the procedure via a VM hook

The now obsolete  `mParseBiggestFloat`  magic and the  `opcParseFloat` 
VM operation are removed.

For csources compatibility, the old magic procedure is still used when
the new   `nimskullHasNoParseFloatMagic`  conditional symbol is absent.
The related  `.compilerprocs`  in  `strmantle`  and  `jssys`  also have
to be kept.