io: only define stdmsg when stdout/err are defined (#1547)
## Summary
Makes sure that `stdmsg` is only defined when `stdout` and `stderr`
are defined.

## Details
For JS/VM backends,  `stderr`  and  `stdout`  doesn't exist, making 
`stdmsg`  unusable. However, feature detection via 
`when declared(stdmsg)`  would not be able to recognize this as 
`stdmsg`  remains defined even in those scenarios.

This PR moves  `stdmsg`  definition to be under the same conditionals
that defines  `stdout` / `stderr` , making sure that it is only defined
when it can be used.