cmdline: new module for building CLIs (#1496)
## Summary

Introducing  `cmdline` , a new experimental module aiming to simplify
the process of building command line interfaces.

## Details

The standard library currently lacks an option for quickly spinning up
command line tools. While  `parseopt`  and  `lexopt`  can handle the
process of reading command line parameters, they are not equipped to:

- Generate help messages for a command
- Parse the command line corresponding to a declared interface
- Automatically handle errors in user input and print helpful
diagnostic

`cmdline`  is designed to fulfill those requirements, and to become the
standard method to construct command line interfaces.

A key part in the design of this module is to facilitate support for
other frontends. Many applications now feature an interface that
provides:

- Colored errors
- Colored help messages
- Producing auto correct suggestions on errors
- etc.

While  `cmdline`  can implement all of these features on its own, it
would balloon the complexity of the module beyond what is reasonable for
a standard library component (factoring in the current maintainer
bandwidth). As such a few design constraints was applied:

- The core structures ( `Cli`  and  `parse()` ) are tightly coupled for
both simplicity and to enforce a standardized interpretation.
- Core structures avoid the use of meta programming: a  `Cli`  must be
able to be built using only  `proc` s. This provide a stable basis for
sugar like DSLs and automatic generators to work with.
- Auxiliary functions like  `help()` ,  `run()`  are implemented using
**only** public interfaces, making sure that user can bring their own
implementation. Additionally, the components that generate portions of
help messages/diagnostics are also made public, allowing a custom
composite to be made.

---------

Co-authored-by: zerbina <100542850+zerbina@users.noreply.github.com>