transf: remove the `env` symbol from the context
Instead of returning an `env` symbol from lambda lifting and then
passing it back into `liftIterSym`, lambda lifting now creates a
preliminary closure construction with the environment var, which is
then transformed into a proper construction later on.

This makes it a little simpler to understand what's going on, and it
also removes a side channel.