lambdalifting: omit unnecessary env creation (#1554)
## Summary

* improve call performance for inner routines that call closure
  routines but don't have closed-over locals themselves
* improve performance in general for closures where some parent
  routine(s) don't have closed-over locals

## Details

Previously, a local closure environment was always allocated on entry
of a routine A that called routines closing over some locals outside of
A. If such routine doesn't have closed-over locals itself, the closure
environment only stored an up-pointer, thus being unnecessary.

An example of such a case:
```nim
proc top() =
  var x = 0
  proc inner() = # <- an closure environment was allocate here
    proc inner2() =
      x = 1
    inner2()
  ...
```

Lambda-lifting is changed to never create a local closure environment
for closure procedures that have no closed-over locals. Instead, the
environment parameter is passed to the inner closures directly.

To make the above change easier, the lambda-lifting pass is changed to
not return the env variable symbol anymore, instead creating a
preliminary `nkClosure` construction with the env symbol that the main
`transf` pass then completes.